Move all dialogs in a dedicated package
