+++ src/main/java/org/mastodon/mamut/collaboration/MastodonGitController.java
@@ -40,7 +40,10 @@ import org.mastodon.mamut.KeyConfigScopes;
 import org.mastodon.mamut.collaboration.commands.MastodonGitCloneRepository;
 import org.mastodon.mamut.collaboration.commands.MastodonGitCreateRepository;
 import org.mastodon.mamut.collaboration.commands.MastodonGitNewBranch;
-import org.mastodon.mamut.collaboration.commands.SetAuthorDialog;
+import org.mastodon.mamut.collaboration.dialogs.SetAuthorDialog;
+import org.mastodon.mamut.collaboration.dialogs.CommitMessageDialog;
+import org.mastodon.mamut.collaboration.dialogs.ErrorDialog;
+import org.mastodon.mamut.collaboration.dialogs.NotificationDialog;
 import org.mastodon.mamut.collaboration.settings.MastodonGitSettingsService;
 import org.mastodon.mamut.collaboration.utils.ActionDescriptions;
 import org.mastodon.mamut.collaboration.utils.BasicDescriptionProvider;
