+++ src/main/java/org/mastodon/mamut/collaboration/commands/MastodonGitCreateRepository.java
@@ -30,7 +30,7 @@ package org.mastodon.mamut.collaboration.commands;
 
 import java.io.File;
 
-import org.mastodon.mamut.collaboration.ErrorDialog;
+import org.mastodon.mamut.collaboration.dialogs.ErrorDialog;
 import org.scijava.ItemVisibility;
 import org.scijava.command.Command;
 import org.scijava.plugin.Parameter;
