+++ src/main/java/org/mastodon/mamut/collaboration/dialogs/ErrorDialog.java
@@ -1,4 +1,4 @@
-package org.mastodon.mamut.collaboration;
+package org.mastodon.mamut.collaboration.dialogs;
 
 import java.awt.Color;
 import java.awt.Dimension;
