+++ src/test/java/org/mastodon/mamut/collaboration/dialogs/CommitMessageDialogDemo.java
@@ -1,4 +1,4 @@
-package org.mastodon.mamut.collaboration;
+package org.mastodon.mamut.collaboration.dialogs;
 
 public class CommitMessageDialogDemo
 {
