+++ src/test/java/org/mastodon/mamut/collaboration/dialogs/NotificationDialogDemo.java
@@ -1,4 +1,4 @@
-package org.mastodon.mamut.collaboration;
+package org.mastodon.mamut.collaboration.dialogs;
 
 public class NotificationDialogDemo
 {
